+++ demisto_sdk/commands/generate_docs/common.py
@@ -273,15 +273,17 @@ def build_example_dict(command_examples: list, insecure: bool):
     examples = {}  # type: dict
     errors = []  # type: list
     for example in command_examples:
-        cmd, md_example, context_example, cmd_errors = execute_command(example, insecure)
+        name, md_example, context_example, cmd_errors = execute_command(example, insecure)
         if 'playbookQuery' in context_example:
             del context_example['playbookQuery']
 
         context_example = json.dumps(context_example, indent=4)
         errors.extend(cmd_errors)
 
-        if not cmd_errors and cmd not in examples:
-            examples[cmd] = (example, md_example, context_example)
+        if not cmd_errors:
+            if name not in examples:
+                examples[name] = []
+            examples[name].append((example, md_example, context_example))
     return examples, errors