+++ demisto_sdk/commands/init/contribution_converter.py
@@ -263,7 +263,7 @@ class ContributionConverter:
             update_docker=True,
             verbose=True,
             assume_yes=True,
-            prev_ver='xsoar-contrib/master'  # default is demisto/master
+            prev_ver='xsoar-contrib/master',  # default is demisto/master
         )
 
     def generate_readme_for_pack_content_item(self, yml_path: str) -> None: